refactor workspace add asset helpers
