+++ .sampo/changesets/issue-359-cli-add-workspace-assets.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split workspace add pattern and binding-source helpers into a focused runtime module without changing the existing CLI surface.
